+++ demo/src/main/java/com/kajucode/patient/controller/PatientController.java
@@ -19,9 +19,9 @@ import com.kajucode.patient.controller.convert.ControllerConverter;
 import com.kajucode.patient.controller.dto.PatientCreationRequest;
 import com.kajucode.patient.controller.dto.PatientResponse;
 import com.kajucode.patient.controller.dto.PatientUpdateRequest;
-import com.kajucode.patient.repository.dto.PatientDto;
 import com.kajucode.patient.repository.entity.PatientEntity;
 import com.kajucode.patient.service.PatientService;
+import com.kajucode.patient.service.dto.PatientDto;
 
 import org.springframework.http.HttpStatus;
