+++ demo/src/main/java/com/kajucode/patient/controller/convert/ControllerConverter.java
@@ -2,8 +2,8 @@ package com.kajucode.patient.controller.convert;
 
 import com.kajucode.patient.controller.dto.PatientResponse;
 import com.kajucode.patient.controller.dto.PatientUpdateRequest;
-import com.kajucode.patient.repository.dto.PatientDto;
 import com.kajucode.patient.repository.entity.PatientEntity;
+import com.kajucode.patient.service.dto.PatientDto;
 
 import lombok.experimental.UtilityClass;
 
@@ -22,6 +22,7 @@ public class ControllerConverter {
 				        			.observations(patientUpdateRequest.getObservations())
 				        			.build();
     }
+	
 	public PatientResponse convertPatientDtoToPatientResponse (PatientDto patientDto) {
     	return PatientResponse.builder().fullName(patientDto.getFullName())
 										.dni(patientDto.getDni())
@@ -36,33 +37,7 @@ public class ControllerConverter {
 										.build();
 
     }
-	public PatientResponse convertPatientEntityToPatientResponse (PatientEntity patientEntity) {
-    	return PatientResponse.builder().fullName(patientEntity.getFullName())
-										.dni(patientEntity.getDni())
-										.age(patientEntity.getAge())
-										.contactNumber(patientEntity.getContactNumber())
-										.address(patientEntity.getAddress())
-										.email(patientEntity.getEmail())
-										.occupation(patientEntity.getOccupation())
-										.dateOfAdmission(patientEntity.getDateOfAdmission())
-										.lifeStory(patientEntity.getLifeStory())
-										.observations(patientEntity.getObservations())
-										.build();
-	}
-    	public PatientDto convertPatientEntityToDtoPatient (PatientEntity patientEntity) {
-            return PatientDto.builder().fullName(patientEntity.getFullName())
-    									.dni(patientEntity.getDni())
-    									.age(patientEntity.getAge())
-    									.contactNumber(patientEntity.getContactNumber())
-    									.address(patientEntity.getAddress())
-    									.email(patientEntity.getEmail())
-    									.occupation(patientEntity.getOccupation())
-    									.dateOfAdmission(patientEntity.getDateOfAdmission())
-    									.lifeStory(patientEntity.getLifeStory())
-    									.observations(patientEntity.getObservations())
-    									.build();
-
-    }
+	
 	public PatientEntity convertToEntityPatient (PatientDto patientDto) {
         PatientEntity patientEntity = new PatientEntity();
         patientEntity.setFullName(patientDto.getFullName());
@@ -77,5 +52,4 @@ public class ControllerConverter {
         patientEntity.setObservations(patientDto.getObservations());
         return patientEntity;
     }
-
 }
