+++ demo/src/main/java/com/kajucode/patient/service/dto/PatientDto.java
@@ -1,4 +1,4 @@
-package com.kajucode.patient.repository.dto;
+package com.kajucode.patient.service.dto;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,8 +6,6 @@ import lombok.Setter;
 
 import java.io.Serializable;
 import java.util.Date;
-
-
 
 @Builder
 @Getter
