--- demo/src/main/java/com/kajucode/patient/service/excepcion/PatientNotFoundException.java
@@ -1,8 +1,0 @@
-package com.kajucode.patient.service.excepcion;
-
-public class PatientNotFoundException extends RuntimeException {
-
-	public PatientNotFoundException(String message) {
-        super(message);
-    }
-}
